+++ graphql-jpa-query-schema/src/test/java/com/introproventures/graphql/jpa/query/schema/GraphQLExecutorTests.java
@@ -19,7 +19,6 @@ package com.introproventures.graphql.jpa.query.schema;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.HashMap;
-import java.util.UUID;
 
 import javax.persistence.EntityManager;
 
@@ -30,6 +29,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.context.annotation.Bean;
+import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.util.Assert;
 
@@ -38,6 +38,7 @@ import com.introproventures.graphql.jpa.query.schema.impl.GraphQLJpaSchemaBuilde
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment=WebEnvironment.NONE)
+@TestPropertySource({"classpath:hibernate.properties"})
 public class GraphQLExecutorTests {
     
     @SpringBootApplication
@@ -57,10 +58,8 @@ public class GraphQLExecutorTests {
         
     }
     
-
     @Autowired
     private GraphQLExecutor executor;
-
     
     @Test
     public void contextLoads() {
@@ -222,6 +221,131 @@ public class GraphQLExecutorTests {
         assertThat(result.toString()).isEqualTo(expected);
     }
 
+    @Test
+    public void queryForEnumIn() {
+        //given
+        String query = "{ Books(where: {genre: {IN: PLAY}}) { select { id title, genre } }}";
+        
+        String expected = "{Books={select=["
+        		+ "{id=5, title=The Cherry Orchard, genre=PLAY}, "
+        		+ "{id=6, title=The Seagull, genre=PLAY}, "
+        		+ "{id=7, title=Three Sisters, genre=PLAY}"
+        		+ "]}}";
+
+        //when
+        Object result = executor.execute(query).getData();
+
+        // then
+        assertThat(result.toString()).isEqualTo(expected);
+    }
+    
+    @Test
+    public void queryForEnumInArray() {
+        //given
+        String query = "{ Books(where: {genre: {IN: [NOVEL, PLAY]}}) { select { id title, genre } }}";
+        
+        String expected = "{Books={select=["
+        		+ "{id=2, title=War and Peace, genre=NOVEL}, "
+        		+ "{id=3, title=Anna Karenina, genre=NOVEL}, "
+        		+ "{id=5, title=The Cherry Orchard, genre=PLAY}, "
+        		+ "{id=6, title=The Seagull, genre=PLAY}, "
+        		+ "{id=7, title=Three Sisters, genre=PLAY}"
+        		+ "]}}";
+
+        //when
+        Object result = executor.execute(query).getData();
+
+        // then
+        assertThat(result.toString()).isEqualTo(expected);
+    }
+
+    @Test
+    public void queryForEnumNinArray() {
+        //given
+        String query = "{ Books(where: {genre: {NIN: [NOVEL]}}) { select { id title, genre } }}";
+        
+        String expected = "{Books={select=["
+        		+ "{id=5, title=The Cherry Orchard, genre=PLAY}, "
+        		+ "{id=6, title=The Seagull, genre=PLAY}, "
+        		+ "{id=7, title=Three Sisters, genre=PLAY}"
+        		+ "]}}";
+
+        //when
+        Object result = executor.execute(query).getData();
+
+        // then
+        assertThat(result.toString()).isEqualTo(expected);
+    }
+    
+    @Test
+    public void queryForEnumEq() {
+        //given
+        String query = "{ Books(where: {genre: {EQ: NOVEL}}) { select { id title, genre } }}";
+        
+        String expected = "{Books={select=["
+        		+ "{id=2, title=War and Peace, genre=NOVEL}, "
+        		+ "{id=3, title=Anna Karenina, genre=NOVEL}"
+        		+ "]}}";
+
+        //when
+        Object result = executor.execute(query).getData();
+
+        // then
+        assertThat(result.toString()).isEqualTo(expected);
+    }
+
+    @Test
+    public void queryForEnumNe() {
+        //given
+        String query = "{ Books(where: {genre: {NE: PLAY}}) { select { id title, genre } }}";
+        
+        String expected = "{Books={select=["
+        		+ "{id=2, title=War and Peace, genre=NOVEL}, "
+        		+ "{id=3, title=Anna Karenina, genre=NOVEL}"
+        		+ "]}}";
+
+        //when
+        Object result = executor.execute(query).getData();
+
+        // then
+        assertThat(result.toString()).isEqualTo(expected);
+    }
+    
+    @Test
+    public void queryForEnumNin() {
+        //given
+        String query = "{ Books(where: {genre: {NIN: PLAY}}) { select { id title, genre } }}";
+        
+        String expected = "{Books={select=["
+        		+ "{id=2, title=War and Peace, genre=NOVEL}, "
+        		+ "{id=3, title=Anna Karenina, genre=NOVEL}"
+        		+ "]}}";
+
+        //when
+        Object result = executor.execute(query).getData();
+
+        // then
+        assertThat(result.toString()).isEqualTo(expected);
+    }
+        
+    @Test
+    public void queryForParentWithEnum() {
+        //given
+        String query = "{ Books { select { id title, author( where: { genre: { EQ: NOVEL } }) { name } } } }";
+        
+        String expected = "{Books={select=["
+        		+ "{id=2, title=War and Peace, author={name=Leo Tolstoy}}, "
+        		+ "{id=3, title=Anna Karenina, author={name=Leo Tolstoy}}"
+        		+ "]}}";
+
+        //when
+        Object result = executor.execute(query).getData();
+
+        // then
+        assertThat(result.toString()).isEqualTo(expected);
+    }
+    
+    
     // https://github.com/introproventures/graphql-jpa-query/issues/30
     @Test
     public void queryForEntityWithMappedSuperclass() {
@@ -251,21 +375,20 @@ public class GraphQLExecutorTests {
         // then
         assertThat(result.toString()).isEqualTo(expected);
     }
-    
-//    // TODO
-//    @Test
-//    public void queryForEntityWithEmeddableTypeAndWhere() {
-//        //given
-//        String query = "{ Boats { id engine(where: { identification: { EQ: \"12345\"}}) { identification } } }";
-//        
-//        String expected = "{Boats={select[id=1, engine={identification=12345}]}}";
-//
-//        //when
-//        Object result = executor.execute(query).getData();
-//
-//        // then
-//        assertThat(result.toString()).isEqualTo(expected);
-//    }
+
+    @Test
+    public void queryForEntityWithEmeddableTypeAndWhere() {
+        //given
+        String query = "{ Boats { select { id engine(where: { identification: { EQ: \"12345\"}}) { identification } } } }";
+
+        String expected = "{Boats={select=[{id=1, engine={identification=12345}}]}}";
+
+        //when
+        Object result = executor.execute(query).getData();
+
+        // then
+        assertThat(result.toString()).isEqualTo(expected);
+    }