+++ graphql-jpa-query-schema/src/test/resources/data.sql
@@ -107,10 +107,10 @@ insert into thing (id, type) values
     ('2D1EBC5B7D2741979CF0E84451C5BBB1', 'Thing1');
     
 -- Books
-insert into author (id, name) values (1, 'Leo Tolstoy');
+insert into author (id, name, genre) values (1, 'Leo Tolstoy', 'NOVEL');
 insert into book (id, title, author_id, genre) values (2, 'War and Peace', 1, 'NOVEL');
 insert into book (id, title, author_id, genre) values (3, 'Anna Karenina', 1, 'NOVEL');
-insert into author (id, name) values (4, 'Anton Chekhov');
+insert into author (id, name, genre) values (4, 'Anton Chekhov', 'PLAY');
 insert into book (id, title, author_id, genre) values (5, 'The Cherry Orchard', 4, 'PLAY');
 insert into book (id, title, author_id, genre) values (6, 'The Seagull', 4, 'PLAY');
 insert into book (id, title, author_id, genre) values (7, 'Three Sisters', 4, 'PLAY');